+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.5.0] - 2022-02-16
 ### Added
 - Reboot requirement info to approval message
+- Ability to set window when update is automatically approved or installed in
+  case of delayed approve
 
 ### Changed
 - Switch from state log to changelog and improve notifications about changes and